+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+[3.0.0]
+* ported to TS
+* rework of the entire lib
+    * Now using `ioredis`
+    * configuration passed as json string and not seperate values
+    * firebase object now handles messaging (if enabled by your plan) as well
+    * user create/get/update/delete
+
 [2.0.0]
 * async initialize (breaking change)